moved RepostoryDetails into CardDetails folder for better structure
